requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
python-dotenv==1.0.1
aiohttp==3.9.5
aiosqlite==0.20.0
pydantic==2.8.2
solana==0.30.2
websockets==12.0
ujson==5.10.0
backoff==2.2.1
